+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for searching projects
+- Support for searching sections
 
 ## [3.1.0] - 2025-05-07